Preparation for moving Docker/test/selenium/specs into test/specs
