+++ Docker/test/selenium/specs/fedprops/item.js
@@ -3,7 +3,7 @@
 const Util = require( 'wdio-mediawiki/Util' );
 const assert = require( 'assert' );
 const WikibaseApi = require( 'wdio-wikibase/wikibase.api' );
-const QueryServiceUI = require( '../../pages/queryservice-ui/queryservice-ui.page' );
+const QueryServiceUI = require( '../../helpers/pages/queryservice-ui/queryservice-ui.page' );
 const ItemPage = require( 'wdio-wikibase/pageobjects/item.page' );
 
 describe( 'Fed props Item', function () {
