+++ Docker/test/selenium/specs/repo/extensions/bananas.lua
