+++ Docker/test/selenium/specs/repo/extensions/entityschema.sx
